✨ feat: member activities 테이블 추가
