--- src/main/java/com/cheeup/domain/PortfolioVeterans.java
@@ -1,27 +1,0 @@
-package com.cheeup.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class PortfolioVeterans {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Integer status;
-
-    @Column(nullable = false, length = 20)
-    private String relation;
-
-    @Column(nullable = false, length = 20)
-    private String number;
-
-    @Column(nullable = false)
-    private Integer ratio;
-}
