+++ src/main/java/com/cheeup/domain/enums/ActivityType.java
@@ -1,0 +1,5 @@
+package com.cheeup.domain.enums;
+
+public enum ActivityType {
+    글_작성;
+}
