+++ src/main/java/com/cheeup/domain/enums/Tier.java
@@ -1,0 +1,5 @@
+package com.cheeup.domain.enums;
+
+public enum Tier {
+    BRONZE, SILVER, GOLD, PLATINUM, DIAMOND, RUBY;
+}
