+++ src/main/java/com/cheeup/domain/job/JobNoticeScraps.java
@@ -1,9 +1,8 @@
-package com.cheeup.domain;
+package com.cheeup.domain.job;
 
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
@@ -14,7 +13,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @EntityListeners(AuditingEntityListener.class)
-public class InactiveMembers {
+public class JobNoticeScraps {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -22,6 +21,4 @@ public class InactiveMembers {
 
     @CreatedDate
     private LocalDateTime createdAt;
-
-    private LocalDateTime deleteAt;
 }
