+++ src/main/java/com/cheeup/domain/member/MemberActivity.java
@@ -1,5 +1,6 @@
 package com.cheeup.domain.member;
 
+import com.cheeup.domain.enums.ActivityType;
 import com.cheeup.domain.enums.MemberRole;
 import com.cheeup.domain.enums.Tier;
 import jakarta.persistence.*;
@@ -16,53 +17,22 @@ import java.time.LocalDateTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @EntityListeners(AuditingEntityListener.class)
-@Table(name = "member")
-public class Member {
-
+@Table(name = "member_activities")
+public class MemberActivity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 30, unique = true)
-    private String kakaoId;
-
-    @Column(nullable = false, length = 20)
-    private String name;
-
-    @Column(nullable = false, length = 50, unique = true)
-    private String email;
-
-    @Column(nullable = false, length = 20, unique = true)
-    private String nickname;
-
-    @Column(nullable = false, length = 20)
     @Enumerated(EnumType.STRING)
-    private MemberRole role = MemberRole.GUEST;
-
-    @Column(length = 225)
-    private String githubLink;
-
-    @Column(length = 225)
-    private String profileImageUrl;
+    @Column(nullable = false)
+    private ActivityType type;
 
     @Column(nullable = false)
-    private Boolean isActivated = true;
-
-    @Column(nullable = false)
-    private Integer experience = 0;
-
-    @Enumerated(EnumType.STRING)
-    private Tier tier = Tier.BRONZE;
+    private Integer experience;
 
     @CreatedDate
     @Column(updatable = false, nullable = false)
     private LocalDateTime createdAt;
 
-    @LastModifiedDate
-    @Column(nullable = false)
-    private LocalDateTime updatedAt;
-
     private LocalDateTime deletedAt;
-
 }
-
