+++ src/main/java/com/cheeup/domain/portfolio/PortfolioCareers.java
@@ -1,27 +1,33 @@
-package com.cheeup.domain;
+package com.cheeup.domain.portfolio;
 
-import com.cheeup.domain.enums.DisabilityType;
+import com.cheeup.domain.enums.CareerStatus;
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.util.Date;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class PortfolioDisabilities {
+public class PortfolioCareers {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
-    private DisabilityType type;
+    private CareerStatus status;
+
+    @Column(nullable = false, length = 20)
+    private String company;
 
     @Column(nullable = false)
-    private Integer grade;
+    private Date startDate;
 
     @Column(nullable = false)
-    private Integer description;
+    private Date endDate;
+
+    private String description;
 }
-
