+++ src/main/java/com/cheeup/domain/portfolio/PortfolioHighSchools.java
@@ -1,5 +1,7 @@
-package com.cheeup.domain;
+package com.cheeup.domain.portfolio;
 
+import com.cheeup.domain.enums.HighSchoolMajor;
+import com.cheeup.domain.enums.HighSchoolType;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -10,21 +12,22 @@ import java.util.Date;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class PortfolioTheses {
+public class PortfolioHighSchools {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 255)
-    private String title;
-
-    private String result;
-
-    @Column(length = 30)
-    private String submissionTarget;
+    @Column(nullable = false, length = 100)
+    private String name;
 
     @Column(nullable = false)
-    private Date submissionDate;
+    @Enumerated(EnumType.STRING)
+    private HighSchoolMajor major;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private HighSchoolType type;
 
     @Column(nullable = false)
     private Date startDate;
