+++ src/main/java/com/cheeup/domain/portfolio/PortfolioMilitaries.java
@@ -1,9 +1,6 @@
-package com.cheeup.domain;
+package com.cheeup.domain.portfolio;
 
-import com.cheeup.domain.enums.GraduateDegree;
-import com.cheeup.domain.enums.UniversityCampus;
-import com.cheeup.domain.enums.UniversityEntry;
-import com.cheeup.domain.enums.UniversityStatus;
+import com.cheeup.domain.enums.*;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -14,37 +11,33 @@ import java.util.Date;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class PortfolioGraduates {
-
+public class PortfolioMilitaries {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 100)
-    private String name;
-
-    @Column(nullable = false, length = 20)
-    private String region;
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private MilitaryType type;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
-    private GraduateDegree degree;
+    private MilitaryBranch branch;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
-    private UniversityCampus campus;
+    private MilitarySpecialty specialty;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
-    private UniversityEntry entry;
+    private MilitaryRank rank;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
-    private UniversityStatus status;
+    private MilitaryDischarge discharge;
 
     @Column(nullable = false)
     private Date startDate;
 
-    @Column(nullable = false)
     private Date endDate;
 }
