+++ src/main/java/com/cheeup/domain/portfolio/PortfolioPersonalInfos.java
@@ -1,27 +1,29 @@
-package com.cheeup.domain;
+package com.cheeup.domain.portfolio;
 
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.util.Date;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class PortfolioCertificates {
-
+public class PortfolioPersonalInfos {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 255)
-    private String title;
+    @Column(nullable = false, length = 10)
+    private String name;
+
+    @Column(nullable = false, length = 50)
+    private String email;
 
     @Column(nullable = false, length = 255)
-    private String rating;
+    private String githubUrl;
 
-    @Column(nullable = false)
-    private Date awardDate;
+    @Column(nullable = false, length = 255)
+    private String blogUrl;
+
+    private String introduction;
 }
