+++ src/main/java/com/cheeup/domain/portfolio/PortfolioSkills.java
@@ -1,27 +1,20 @@
-package com.cheeup.domain;
+package com.cheeup.domain.portfolio;
 
+import com.cheeup.domain.enums.SkillLevel;
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.util.Date;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class PortfolioAwards {
+public class PortfolioSkills {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 255)
-    private String title;
-
-    @Column(nullable = false, length = 255)
-    private String rating;
-
     @Column(nullable = false)
-    private Date acquisitionDate;
-
+    @Enumerated(EnumType.STRING)
+    private SkillLevel level;
 }
